Replace UnityPlayer with UnityPlayerForActivityOrService. Handle UnityPlayer no longer extending a FrameLayout.
