+++ flutter_embed_unity_6000_android/android/src/main/kotlin/com/learntoflutter/flutter_embed_unity_android/platformView/CopyMotionEvent.kt
@@ -1,6 +1,6 @@
 // source https://gist.github.com/sebschaef/b803da53217c88e8c691aeed08602193
 
-package com.learntoflutter.flutter_embed_unity_android.unity
+package com.learntoflutter.flutter_embed_unity_android.platformView
 
 import android.view.MotionEvent
