+++ flutter_embed_unity_6000_android/android/src/main/kotlin/com/learntoflutter/flutter_embed_unity_android/unity/UnityPlayerSingleton.kt
@@ -6,12 +6,12 @@ import android.view.InputDevice
 import android.view.MotionEvent
 import android.view.View
 import com.learntoflutter.flutter_embed_unity_android.constants.FlutterEmbedConstants.Companion.logTag
-import com.unity3d.player.UnityPlayer
+import com.unity3d.player.UnityPlayerForActivityOrService
 import io.flutter.Log
 
 
 @SuppressLint("ViewConstructor")
-class UnityPlayerSingleton private constructor (activity: Activity) : UnityPlayer(activity) {
+class UnityPlayerSingleton private constructor (activity: Activity) : UnityPlayerForActivityOrService(activity) {
     companion object {
         // We must use a singleton UnityPlayer, because it was never designed to be
         // reused in multiple views. Calling unityPlayer.destroy() will kill the
@@ -74,49 +74,6 @@ class UnityPlayerSingleton private constructor (activity: Activity) : UnityPlaye
         fun getInstance(): UnityPlayerSingleton? {
             return singleton
         }
-    }
-
-    // This is required for Unity to receive touch events
-    @SuppressLint("ClickableViewAccessibility")
-    override fun onTouchEvent(motionEvent: MotionEvent): Boolean {
-        motionEvent.source = InputDevice.SOURCE_TOUCHSCREEN
-//        Log.i(logTag, "onTouchEvent")
-    
-         // true for Flutter Virtual Display, false for Hybrid composition.
-        if (motionEvent.deviceId == 0) {        
-            /* 
-              Flutter creates a touchscreen motion event with deviceId 0. (https://github.com/flutter/flutter/blob/34b454f42dd6f8721dfe43fc7de5d215705b5e52/packages/flutter/lib/src/services/platform_views.dart#L639)
-              Unity's new Input System package does not detect these touches, copy the motion event to change the immutable deviceId.
-            */
-            val modifiedEvent = motionEvent.copy(deviceId = -1)
-            motionEvent.recycle()
-            return super.onTouchEvent(modifiedEvent)
-        } else {
-            return super.onTouchEvent(motionEvent)
-        }
-    }
-
-    override fun onWindowVisibilityChanged(visibility: Int) {
-        Log.d(logTag, "UnityPlayerSingleton onWindowVisibilityChanged $visibility")
-
-        if(visibility == View.VISIBLE) {
-            // For some unknown reason, if window visibility changes quickly from View.VISIBLE
-            // to View.GONE and back to View.VISIBLE, Unity UI appears to freeze.
-            // This happens, for example, on orientation change, flutter hot reload, and
-            // occasionally on widget rebuild if there is a significant change to the widget
-            // tree (you can usually see this as a brief flicker of the widget).
-            // The underlying UnityPlayer is still active and still responds to messages even
-            // though it appears frozen, so it is purely a UI thing. Presumably a bug in Unity.
-            // However using UnityPlayer to render in a View like this is not supported so
-            // unlikely to be fixed by Unity
-            // (see https://docs.unity3d.com/Manual/UnityasaLibrary-Android.html)
-            // As a workaround, pause and resume the player unfreezes the UI
-            Log.d(logTag, "UnityPlayerSingleton became visible, so pausing and resuming Unity")
-            pause()
-            resume()
-        }
-
-        super.onWindowVisibilityChanged(visibility)
     }
 
     // Overriding kill() was an experiment to try to resolve app closing / crashing when
